[sarlc] Path detector is now injectable.
