+++ products/sarlc/src/main/java/io/sarl/lang/sarlc/commands/PathDetector.java
@@ -1,0 +1,88 @@
+/*
+ * $Id$
+ *
+ * SARL is an general-purpose agent programming language.
+ * More details on http://www.sarl.io
+ *
+ * Copyright (C) 2014-2018 the original authors or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.sarl.lang.sarlc.commands;
+
+import java.io.File;
+import java.util.List;
+
+import com.google.inject.ImplementedBy;
+
+/**
+ * Detect the three base paths when they are missed.
+ * The paths are: <ul>
+ * <li>the SARL output path,</li>
+ * <li>the Java class output path, and</li>
+ * <li>the working path.</li>
+ * </ul>
+ *
+ * @author $Author: sgalland$
+ * @version $FullVersion$
+ * @mavengroupid $GroupId$
+ * @mavenartifactid $ArtifactId$
+ * @since 0.8
+ */
+@ImplementedBy(DefaultPathDetector.class)
+public interface PathDetector {
+
+	/** Change the SARL output path.
+	 *
+	 * @param path the path.
+	 */
+	void setSarlOutputPath(File path);
+
+	/** Change the Java class output path.
+	 *
+	 * @param path the path.
+	 */
+	void setClassOutputPath(File path);
+
+	/** Change the working path.
+	 *
+	 * @param path the path.
+	 */
+	void setWorkingPath(File path);
+
+	/** Replies the SARL output path.
+	 *
+	 * @return the path.
+	 */
+	File getSarlOutputPath();
+
+	/** Replies the Java class output path.
+	 *
+	 * @return the path.
+	 */
+	File getClassOutputPath();
+
+	/** Replies the working path.
+	 *
+	 * @return the path.
+	 */
+	File getWorkingPath();
+
+	/** Resolve the paths.
+	 *
+	 * @param args the command line arguments to consider.
+	 */
+	void resolve(List<String> args);
+
+}
